Use hatch as build system
