+++ .pre-commit-config.yaml
@@ -23,7 +23,3 @@ repos:
   hooks:
     - id: pyupgrade
       args: ["--py39-plus"]
-- repo: https://github.com/miki725/pre-commit-twine-check
-  rev: '0.1'
-  hooks:
-  - id: twine-check
