--- MANIFEST.in
@@ -1,11 +1,0 @@
-include LICENSE
-include README.rst
-graft tests
-include conftest.py
-include pytest.ini
-include test_requirements.txt
-include override_requirements.txt
-include tox.ini
-include src/libertem_live/detectors/LICENSE
-recursive-exclude **/.mypy_cache *
-recursive-exclude **/__pycache__ *
